+++ extensions/easydict/src/types.ts
@@ -2,7 +2,7 @@
  * @author: tisfeng
  * @createTime: 2022-06-04 21:58
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-08-19 16:51
+ * @lastEditTime: 2022-08-20 11:13
  * @fileName: types.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
@@ -21,6 +21,7 @@ import { LanguageItem } from "./language/type";
 export interface ActionListPanelProps {
   displayItem: ListDisplayItem;
   isInstalledEudic: boolean;
+  isShowingReleasePrompt: boolean;
   onLanguageUpdate: (language: LanguageItem) => void;
 }